refactor(paper): full split — move manuscript.py -> paper/stats.py
- git mv src/toxfam/evaluation/manuscript.py -> paper/stats.py; paper stats now
  leave the installable wheel (verified: 0 paper refs in the wheel, manuscript.py
  absent, 32 toxfam/ modules still packaged)
- git mv tests/test_manuscript.py -> paper/tests/test_stats.py
- rewrite 5 figure modules + the test: toxfam.evaluation.manuscript -> paper.stats
- paper/stats.py keeps its toxfam.evaluation.{hbi,metrics} imports (paper->toxfam,
  one-directional). evaluation/__init__ never exported manuscript -> no other change.

Suite: 157 passed; toxfam --help intact.
