+++ paper/stats.py
